Round and Format Quantiles

Round the values of the quantiles and confidence limits and select those corresponding to "round" exceedance probabilities.

Usage

# S3 method for quantile.pgpTList
format(
  x,
  digits = 2,
  probExc = as.vector(outer(c(5, 2, 1), c(0.1, 0.01, 0.001, 1e-04))),
  ...
)

Arguments

x

An object with class "quantile.pgpgTList" as computed by the quantile method of the "pgpgTList" class.

digits

The number of digits to be used for the colums Quant, L and U columns representing the quantile and the confidence limits.

probExc

A numeric vector giving the exceedances probabilities to be displayed among those available in x. The default values correspond to 12 values 0.5, 0.2, 0.1, ..., 1e-4.

...

Not used.

Value

A data frame with the rounded quantiles and confidence limits at the" chosen "round" exceedance probabilities.

Details

It is often needed to show a table of quantiles corresponding to round exceedance probabilities such as 0.01, 0.001. Moreover the quantiles and confidence limits as printed by default with number of digits which may be higher than needed. Most often, a number of digits between 0 and 2 is hight enough in relation with the precision of the data used to compute the quantiles.