Compute an Estimate of the Tail Coefficient 'xi' using Quantile Regression Results

Estimate the tail coefficient \(\xi\) from quantiles.

Usage

# S3 method for rqTList
xi(object, tau = NULL, lastFullYear = TRUE, plot = FALSE, ...)

Arguments

object

An object with class "rqTList".

tau

A vector of 3 probabilities. This vector must be in strictly increasing order and contain values that are found in tau(object). By default the three largest values in tau(object) are used.

lastFullYear

Logical. It TRUE the value of the estimated xi is computed for the last full year of the data used to create data. See predict.rqTList.

plot

Logical. If TRUE a (gg)plot is shown.

...

Not used yet.

Value

A data frame with a xiHat column containing the estimation.

Details

Let \(\tau_1 < \tau_2 < \tau_3\) be the probabilities and \(T_1 < T_2 < T_3\) be the corresponding return periods \(T_i = 1 / (1 - \tau_i)\). Let \(q_1\), \(q_2\), \(q_3\) be the corresponding quantiles as computed by quantile regression. Then we can find \(\xi\) by solving the equation $$\frac{T_3^\xi - T_1^\xi}{T_2^\xi - T_1^\xi} = \frac{q_3 - q_1}{q_2 - q_1}.$$ The value of \(\xi\) then depends on the covariates used in the quantile regression. We only consider here the case where the quantiles depend on the date, with emphasis on the case where the quantile are periodic functions of the date with one-year period. The the value of \(\xi\) has also the one-year periodicity, and we can assess its variation on a one-year period which can be the last full year in the period used to estimate the quantiles.

Caution

There seems to be a huge uncertainty on this estimate, and it seems that the variations are somewhat exaggerated. An alternative method is estimating \(\xi\) by using a moving time window in the year. Also the question arises as whether declustering should be taking into account.